Move tests to test_jsondiff.py to aid pytest discovery
pytest defaults to running tests from files named `test_*`, so move
the tests into one.  Leave an empty `__init__.py` to keep relative
imports working.
